+++ Source/Plugins/Processing/ProcessingFilters/NeighborCICorrelation.cpp
@@ -230,7 +230,7 @@ void NeighborCICorrelation::execute()
       }
     }
     QList<QString> voxelArrayNames = m->getAttributeMatrix(getCellAttributeMatrixName())->getAttributeArrayNameList();
-    for (size_t j = 0; j < totalPoints; j++)
+    for (int64_t j = 0; j < totalPoints; j++)
     {
       neighbor = bestNeighbor[j];
       if (neighbor != -1)
